requirements.txt
amqp==2.6.1
asgiref==3.6.0
astroid==2.13.2
autopep8==2.0.1
billiard==3.6.4.0
celery==4.4.7
click==7.1
dill==0.3.6
Django==3.1.14
django-autoslug==1.9.8
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
isort==5.11.4
kombu==4.6.11
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.15.10
pytz==2022.7
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
vine==1.3.0
wcwidth==0.2.6
wrapt==1.14.1
